- updated to include vulkan api so i can add vulkan in the future - switched to old meshing system again - new terrain generator

+++ Minecraft Clone/src/RenderEngine/OpenGL/Texture/Types/TextureArray.h
@@ -34,7 +34,7 @@ public:
 	bool AddTextureToArray(RawTextureData* Data) {
 		Format = GL_RGBA;
 		if (Data->data) {
-			if (width == Data->width && height == Data->height) {
+			if (width == (GLuint)Data->width && height == (GLuint)Data->height) {
 				if (Data->format == GL_RGB) {
 					for (int index = 0; index < Data->width * Data->height; index++) {
 						ArrayData.push_back(Data->data[(index * 3)]);
